+++ lsp-types/quickcheck/Language/LSP/Protocol/QuickCheck.hs
@@ -1,0 +1,6 @@
+-- | Provides 'Arbitrary' instances for protocol types.
+module Language.LSP.Protocol.QuickCheck (module Export) where
+
+import Language.LSP.Protocol.QuickCheck.Common as Export
+import Language.LSP.Protocol.QuickCheck.Message as Export
+import Language.LSP.Protocol.QuickCheck.Types as Export