+++ lsp-types/src/Language/LSP/Protocol/Meta.hs
@@ -1,0 +1,3 @@
+module Language.LSP.Protocol.Meta (module Export) where
+
+import Language.LSP.Protocol.Internal.Meta as Export